feat: add admin excursions page
