+++ frontend/src/app/admin/page.tsx
@@ -1,187 +1,52 @@
 'use client';
 
-import { useEffect, useState } from 'react';
-import { getUsers, updateUserRole } from '@/services/userService';
-import { useSession } from 'next-auth/react';
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
-  Paper,
-  TablePagination,
-  TextField,
-  MenuItem,
-  Button,
-} from '@mui/material';
+import Link from 'next/link';
+import PeopleIcon from '@mui/icons-material/People';
+import TourIcon from '@mui/icons-material/Tour';
 
-interface User {
-  id: string;
-  firstName: string;
-  lastName: string;
-  email: string;
-  role: string;
-}
-
-export default function AdminPage() {
-  const { data: session } = useSession();
-
-  const [users, setUsers] = useState<User[]>([]);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  const [filters, setFilters] = useState({
-    name: '',
-    surname: '',
-    role: '',
-    page: 0,
-    pageSize: 5,
-  });
-
-  // const [totalUsers, setTotalUsers] = useState(0);
-
-  const loadUsers = async () => {
-    setLoading(true);
-    setError(null);
-    try {
-      const params = {
-        name: filters.name,
-        surname: filters.surname,
-        role: filters.role,
-        page: filters.page + 1,
-        pageSize: filters.pageSize,
-      };
-      const data = await getUsers(session?.accessToken as string, params);
-      setUsers(data || []);
-      // setTotalUsers(data.total || 0);
-    } catch (err) {
-      setError('Failed to fetch users');
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    loadUsers();
-  }, [filters]);
-
-  // Handle Role Update
-  const handleUpdateRole = async (userId: string, newRole: string) => {
-    try {
-      await updateUserRole(session?.accessToken as string, userId, newRole);
-      loadUsers(); // Refresh the list after role update
-    } catch (err) {
-      setError('Failed to update user role');
-    }
-  };
-
-  // Handle Pagination
-  const handlePageChange = (event: unknown, newPage: number) => {
-    setFilters((prev) => ({ ...prev, page: newPage }));
-  };
-
-  const handlePageSizeChange = (
-    event: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-  ) => {
-    setFilters((prev) => ({
-      ...prev,
-      pageSize: parseInt(event.target.value, 10),
-      page: 0,
-    }));
-  };
-
+export default function AdminPanel() {
   return (
-    <div className="py-6 px-20">
-      <h1 className="text-2xl font-bold mb-4">User Management</h1>
+    <div className="min-h-screen bg-gray-100 flex items-center justify-center p-8">
+      <div className="max-w-4xl w-full bg-white p-8 rounded-lg shadow-lg text-center">
+        {/* Admin Panel Heading */}
+        <h1 className="text-4xl font-bold mb-4">Admin Panel</h1>
+        <p className="text-gray-600 mb-8">
+          Manage users and excursions efficiently.
+        </p>
 
-      {/* Filters */}
-      <div className="mb-4 flex gap-4">
-        <TextField
-          label="First Name"
-          variant="outlined"
-          value={filters.name}
-          onChange={(e) => setFilters({ ...filters, name: e.target.value })}
-          size="small"
-        />
-        <TextField
-          label="Surname"
-          variant="outlined"
-          value={filters.surname}
-          onChange={(e) => setFilters({ ...filters, surname: e.target.value })}
-          size="small"
-        />
-        <TextField
-          label="Role"
-          select
-          value={filters.role}
-          onChange={(e) => setFilters({ ...filters, role: e.target.value })}
-          size="small"
-          className="w-40"
-          variant="outlined"
-        >
-          <MenuItem value="">All</MenuItem>
-          <MenuItem value="User">User</MenuItem>
-          <MenuItem value="Banned">Banned</MenuItem>
-        </TextField>
+        {/* Navigation Links */}
+        <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
+          {/* User Management Link */}
+          <Link href="/admin/users" className="group block">
+            <div className="bg-blue-50 border border-blue-200 rounded-lg p-6 shadow-md hover:shadow-lg transition-transform transform hover:-translate-y-1">
+              <div className="flex flex-col items-center justify-center gap-4">
+                <PeopleIcon className="text-blue-500" style={{ fontSize: 64 }} />
+                <h2 className="text-2xl font-semibold text-gray-800 group-hover:text-blue-600 transition-colors">
+                  User Management
+                </h2>
+                <p className="text-gray-500 text-center">
+                  View and manage user roles, including banning and unbanning users.
+                </p>
+              </div>
+            </div>
+          </Link>
+
+          {/* Excursion Management Link */}
+          <Link href="/admin/excursions" className="group block">
+            <div className="bg-green-50 border border-green-200 rounded-lg p-6 shadow-md hover:shadow-lg transition-transform transform hover:-translate-y-1">
+              <div className="flex flex-col items-center justify-center gap-4">
+                <TourIcon className="text-green-500" style={{ fontSize: 64 }} />
+                <h2 className="text-2xl font-semibold text-gray-800 group-hover:text-green-600 transition-colors">
+                  Excursion Management
+                </h2>
+                <p className="text-gray-500 text-center">
+                  Edit, delete, or manage any excursions available on the platform.
+                </p>
+              </div>
+            </div>
+          </Link>
+        </div>
       </div>
-
-      <TableContainer component={Paper}>
-        <Table>
-          <TableHead>
-            <TableRow>
-              <TableCell>Name</TableCell>
-              <TableCell>Email</TableCell>
-              <TableCell>Role</TableCell>
-              <TableCell>Actions</TableCell>
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {users.map((user) => (
-              <TableRow key={user.id}>
-                <TableCell>
-                  {user.firstName} {user.lastName}
-                </TableCell>
-                <TableCell>{user.email}</TableCell>
-                <TableCell>{user.role}</TableCell>
-                <TableCell>
-                  <Button
-                    variant="contained"
-                    color={user.role === 'User' ? 'error' : 'success'}
-                    size="small"
-                    onClick={() =>
-                      handleUpdateRole(
-                        user.id,
-                        user.role === 'User' ? 'Banned' : 'User'
-                      )
-                    }
-                  >
-                    {user.role === 'User' ? 'Ban' : 'Unban'}
-                  </Button>
-                </TableCell>
-              </TableRow>
-            ))}
-            {!users.length && (
-              <TableRow>
-                <TableCell colSpan={4} align="center">
-                  No users found.
-                </TableCell>
-              </TableRow>
-            )}
-          </TableBody>
-        </Table>
-      </TableContainer>
-
-      <TablePagination
-        component="div"
-        // count={totalUsers}
-        page={filters.page}
-        onPageChange={handlePageChange}
-        rowsPerPage={filters.pageSize}
-        onRowsPerPageChange={handlePageSizeChange}
-        rowsPerPageOptions={[5, 10, 20]}
-      />
     </div>
   );
 }
